# Addresses performance for large selector objects