--- a/src/index.js
+++ b/src/index.js
@@ -32,16 +32,29 @@ export const resolveSelectors = obj => {
   // flag for checking if we have *any*
   let hasAtLeastOneResolved = false
 
+  // keep track of not yet extracted selectors and save significant
+  // loops on large objects
+  const selectorsToExtract = Object.assign({}, obj)
+
   // extract all deps and any resolved items
-  for (const selectorName in obj) {
-    const fn = obj[selectorName]
+  for (const selectorName in selectorsToExtract) {
+    const fn = selectorsToExtract[selectorName]
+
     if (!isResolved(selectorName)) {
       fn.deps = fn.deps.map((val, index) => {
         // if it is a function not a string
         if (val.call) {
+          // if it has been extracted in in earlier calls, exit
+          if (val.selectorName) return val.selectorName
+
           // look for it already on the object
-          for (const key in obj) {
-            if (obj[key] === val) {
+          for (const key in selectorsToExtract) {
+            if (selectorsToExtract[key] === val) {
+              // set selectorName to optimize consecutive calls and remove
+              // from object to not loop through it again
+              val.selectorName = key
+              delete selectorsToExtract[key]
+
               // return its name if found
               return key
             }
@@ -56,7 +69,7 @@ export const resolveSelectors = obj => {
 
         // the `val` is a string that exists on the object return the string
         // we'll resolve it later
-        if (obj[val]) return val
+        if (selectorsToExtract[val]) return val
 
         // if we get here, its a string that doesn't exist on the object
         // which won't work, so we throw a helpful error
@@ -84,7 +97,13 @@ export const resolveSelectors = obj => {
       if (!isResolved(selectorName)) {
         hasUnresolved = true
         if (depsAreResolved(fn.deps)) {
-          obj[selectorName] = fn(obj, fn.deps)
+          // we could just use `obj[selectorName] = fn(obj, fn.deps)`, but that
+          // has a significant performance impact when trying to perform this
+          // on a large object (> 1000). More on this here:
+          // http://2ality.com/2014/01/object-assign.html
+          const selectorFn = fn(obj, fn.deps)
+          delete obj[selectorName]
+          obj[selectorName] = selectorFn
         }
       }
     }