Hi Henrik,

I was talking to @layflags yesterday about the extraction / resolve performance regarding #3 

I had a look into it and found 2 spots that can greatly reduce the time for the `resolveSelectors` function for large selector objects. I made the changes in-line on purpose without refactoring, so we can see the affected areas better in the diff. Happy to clean it up if the PR is something you would be interested in.

I changed 2 main areas:
- the loop to extract selectors
- the resolve loop
- added tests for that also

I did a few benchmarks based on the test cases included in this PR:

| Scenarios | 1.000 selectors  | 10.000 selectors | Comment
| - | - | - | - |
| original | 4.061 ms | 25.478 ms |
| extraction optimization only | 4.020ms | 6.120 ms | Seems to have little effect on 1k selectors |
| resolve optimization only | 63ms | 24.924 ms | Seems to have little effect on 10k selectors |
| both optimizations | 50ms | 6.152 ms |

My setup: 2016 Mac (Mojave), 3.3 GHz Intel Core i7

The main performance improvement is this:
```javascript
// before
obj[selectorName] = fn(obj, fn.deps)

// after
const selectorFn = fn(obj, fn.deps)
delete obj[selectorName]
obj[selectorName] = selectorFn
```
I drilled down that:
- `fn(obj, fn.deps)` does not cost anything (almost always 0ms)
- `obj[selectorName] = fn(obj, fn.deps)` costs an average of 6ms

Happy to receive feedback on it