@rudionrails 2019-03-14
Thanks for the swift reply. @layflags and I are currently looking at this again to make sure this is not just something which works in node, but also Chrome, FF, etc.