@layflags 2019-03-14
OMG, by just applying the `delete` optimization my App works now and I could even reduce the timeout to 20ms 😱