@layflags 2019-03-16
@HenrikJoreteg What about just applying the `delete` optimization for now and publish a patch for this package and `redux-bundler`?