@HenrikJoreteg 2019-03-17
@layflags yeah I think you're right. For my own piece of mind I'm going to try this in my app first just to make sure I'm not missing anything.