@HenrikJoreteg 2019-03-17
@rudionrails @layflags I actually went as far as publishing `4.0.2` before realizing there was a problem, but I unpublished it until we can figure it out. I know it's this line that did it `delete selectorsToExtract[key]` because when i comment that out, all my selectors resolve again.