@rudionrails 2019-03-17
@HenrikJoreteg OK, let me change the PR to only have that `delete obj[selectorName]` optimization. @layflags successfully tested that one in his app.